More CUDA debugging and profiling. Big boost in performance after switching to static local vectors instead of dynamic ones. Seems to be bug free.

+++ gpu_prepare.c
@@ -3,10 +3,10 @@
 int gpu_prepare(int N_data, int N_filters)
 {
 
-    ERR(cudaMalloc(&d_chi2_min, N_BLOCKS * sizeof(double)));
+    ERR(cudaMalloc(&d_chi2_min, N_BLOCKS * sizeof(float)));
     ERR(cudaMalloc(&d_iloc_min, N_BLOCKS * sizeof(long int)));
 
-    ERR(cudaMallocHost(&h_chi2_min, N_BLOCKS * sizeof(double)));
+    ERR(cudaMallocHost(&h_chi2_min, N_BLOCKS * sizeof(float)));
     ERR(cudaMallocHost(&h_iloc_min, N_BLOCKS * sizeof(long int)));
     
     ERR(cudaMallocHost(&dData, N_data * sizeof(struct obs_data)));
